# Conflicts:
#	Controller/src/main/resources/messages_ar.properties
#	Controller/src/main/resources/messages_cn.properties
#	Controller/src/main/resources/messages_en.properties
#	Controller/src/main/resources/messages_in.properties
#	Controller/src/main/resources/messages_ru.properties
#	Dao/src/main/java/me/exrates/dao/impl/OrderDaoImpl.java
#	Service/src/main/java/me/exrates/service/impl/BotServiceImpl.java
#	Service/src/main/java/me/exrates/service/impl/OrderServiceImpl.java

+++ Service/src/main/java/me/exrates/service/impl/OrderServiceImpl.java
@@ -18,11 +18,9 @@ import me.exrates.model.dto.onlineTableDto.OrderListDto;
 import me.exrates.model.dto.onlineTableDto.OrderWideListDto;
 import me.exrates.model.enums.*;
 import me.exrates.model.util.BigDecimalProcessing;
-import me.exrates.model.vo.BackDealInterval;
-import me.exrates.model.vo.CacheData;
-import me.exrates.model.vo.TransactionDescription;
-import me.exrates.model.vo.WalletOperationData;
+import me.exrates.model.vo.*;
 import me.exrates.service.*;
+import me.exrates.service.cache.OrdersStatisticByPairsCache;
 import me.exrates.service.events.AcceptOrderEvent;
 import me.exrates.service.events.CancelOrderEvent;
 import me.exrates.service.events.CreateOrderEvent;
@@ -50,6 +48,7 @@ import java.math.BigDecimal;
 import java.security.Principal;
 import java.time.LocalDateTime;
 import java.util.*;
+import java.util.function.Function;
 import java.util.stream.Collectors;
 
 import static java.util.stream.Collectors.toList;
@@ -59,12 +58,17 @@ import static me.exrates.model.enums.OrderActionEnum.*;
 @Service
 public class OrderServiceImpl implements OrderService {
 
+  public static final String SCOPE = "ALL";
   private static final Logger logger = LogManager.getLogger(OrderServiceImpl.class);
 
   private final BigDecimal MAX_ORDER_VALUE = new BigDecimal(100000);
   private final BigDecimal MIN_ORDER_VALUE = new BigDecimal(0.000000001);
 
   private final BigDecimal RATE_CHECK_DEVIATION_PERCENTS = new BigDecimal(10);
+
+  private final List<BackDealInterval> intervals = Arrays.stream(ChartPeriodsEnum.values())
+          .map(ChartPeriodsEnum::getBackDealInterval)
+          .collect(Collectors.toList());
 
   @Autowired
   private OrderDao orderDao;
@@ -110,28 +114,33 @@ public class OrderServiceImpl implements OrderService {
   @Autowired
   private UserRoleService userRoleService;
   @Autowired
-  private BotService botService;
-  @Autowired
   private ObjectMapper objectMapper;
   @Autowired
   private ApplicationEventPublisher eventPublisher;
+  @Autowired
+  private OrdersStatisticByPairsCache ordersStatisticByPairsCache;
+
+
+  @Override
+  public List<BackDealInterval> getIntervals() {
+    return intervals;
+  }
 
   @Transactional
   @Override
   public ExOrderStatisticsDto getOrderStatistic(CurrencyPair currencyPair, BackDealInterval backDealInterval, Locale locale) {
     ExOrderStatisticsDto result = serviceCacheableProxy.getOrderStatistic(currencyPair, backDealInterval);
     result = new ExOrderStatisticsDto(result);
-    result.setPercentChange(BigDecimalProcessing.formatLocale(BigDecimalProcessing.doAction(
-            result.getFirstOrderRate(), result.getLastOrderRate(), ActionType.PERCENT_GROWTH),
-            locale, 2));
-    result.setFirstOrderAmountBase(BigDecimalProcessing.formatLocale(result.getFirstOrderAmountBase(),locale, true));
-    result.setFirstOrderRate(BigDecimalProcessing.formatLocale(result.getFirstOrderRate(), locale, true));
-    result.setLastOrderAmountBase(BigDecimalProcessing.formatLocale(result.getLastOrderAmountBase(), locale, true));
-    result.setLastOrderRate(BigDecimalProcessing.formatLocale(result.getLastOrderRate(), locale, true));
-    result.setMinRate(BigDecimalProcessing.formatLocale(result.getMinRate(), locale, true));
-    result.setMaxRate(BigDecimalProcessing.formatLocale(result.getMaxRate(), locale, true));
-    result.setSumBase(BigDecimalProcessing.formatLocale(result.getSumBase(), locale, true));
-    result.setSumConvert(BigDecimalProcessing.formatLocale(result.getSumConvert(), locale, true));
+    result.setPercentChange(BigDecimalProcessing.formatNonePoint(BigDecimalProcessing.doAction(
+            result.getFirstOrderRate(), result.getLastOrderRate(), ActionType.PERCENT_GROWTH), 2));
+    result.setFirstOrderAmountBase(BigDecimalProcessing.formatNonePoint(result.getFirstOrderAmountBase(), true));
+    result.setFirstOrderRate(BigDecimalProcessing.formatNonePoint(result.getFirstOrderRate(), true));
+    result.setLastOrderAmountBase(BigDecimalProcessing.formatNonePoint(result.getLastOrderAmountBase(), true));
+    result.setLastOrderRate(BigDecimalProcessing.formatNonePoint(result.getLastOrderRate(), true));
+    result.setMinRate(BigDecimalProcessing.formatNonePoint(result.getMinRate(), true));
+    result.setMaxRate(BigDecimalProcessing.formatNonePoint(result.getMaxRate(), true));
+    result.setSumBase(BigDecimalProcessing.formatNonePoint(result.getSumBase(), true));
+    result.setSumConvert(BigDecimalProcessing.formatNonePoint(result.getSumConvert(), true));
     return result;
   }
 
@@ -184,7 +193,7 @@ public class OrderServiceImpl implements OrderService {
    @Override
    public List<ExOrderStatisticsShortByPairsDto> getOrdersStatisticByPairsEx() {
      Locale locale = Locale.ENGLISH;
-     List<ExOrderStatisticsShortByPairsDto> result = orderDao.getOrderStatisticByPairs();
+     List<ExOrderStatisticsShortByPairsDto> result = ordersStatisticByPairsCache.getCachedList();
      result = result.stream()
              .map(ExOrderStatisticsShortByPairsDto::new)
              .collect(toList());
@@ -222,7 +231,7 @@ public class OrderServiceImpl implements OrderService {
   @Transactional
   @Override
   public List<ExOrderStatisticsShortByPairsDto> getOrdersStatisticByPairsSessionless(Locale locale) {
-    List<ExOrderStatisticsShortByPairsDto> result = orderDao.getOrderStatisticByPairs();
+    List<ExOrderStatisticsShortByPairsDto> result = ordersStatisticByPairsCache.getCachedList();
     result.forEach(e -> {
           BigDecimal lastRate = new BigDecimal(e.getLastOrderRate());
           BigDecimal predLastRate = e.getPredLastOrderRate() == null ? lastRate : new BigDecimal(e.getPredLastOrderRate());
@@ -481,6 +490,15 @@ public class OrderServiceImpl implements OrderService {
     }
   }
 
+  @Override
+  @Transactional
+  public void postBotOrderToDb(OrderCreateDto orderCreateDto) {
+    ExOrder exOrder = new ExOrder(orderCreateDto);
+    exOrder.setUserAcceptorId(orderCreateDto.getUserId());
+    orderDao.postAcceptedOrderToDB(exOrder);
+    eventPublisher.publishEvent(new AcceptOrderEvent(exOrder));
+  }
+
 
   @Override
   @Transactional(rollbackFor = {Exception.class})
@@ -598,7 +616,7 @@ public class OrderServiceImpl implements OrderService {
                                                      String email, CurrencyPair currencyPair, OrderStatus status,
                                                      OperationType operationType,
                                                      String scope, Integer offset, Integer limit, Locale locale) {
-    List<OrderWideListDto> result = orderDao.getMyOrdersWithState(email, currencyPair, status, operationType, scope, offset, limit, locale);
+    List<OrderWideListDto> result = orderDao.getMyOrdersWithState(userService.getIdByEmail(email), currencyPair, status, operationType, scope, offset, limit, locale);
     if (Cache.checkCache(cacheData, result)) {
       result = new ArrayList<OrderWideListDto>() {{
         add(new OrderWideListDto(false));
@@ -1048,6 +1066,12 @@ public class OrderServiceImpl implements OrderService {
   @Override
   public Integer deleteOrderByAdmin(int orderId) {
     OrderCreateDto order = orderDao.getMyOrderById(orderId);
+    OrderRoleInfoForDelete orderRoleInfo = orderDao.getOrderRoleInfo(orderId);
+    if (orderRoleInfo.mayDeleteWithoutProcessingTransactions()) {
+      setStatus(orderId, OrderStatus.DELETED);
+      return 1;
+    }
+
     Object result = deleteOrder(orderId, OrderStatus.DELETED, DELETE);
     if (result instanceof OrderDeleteStatus) {
       if ((OrderDeleteStatus) result == OrderDeleteStatus.NOT_FOUND) {
@@ -1296,7 +1320,7 @@ public class OrderServiceImpl implements OrderService {
   public List<OrderWideListDto> getUsersOrdersWithStateForAdmin(String email, CurrencyPair currencyPair, OrderStatus status,
                                                                 OperationType operationType,
                                                                 Integer offset, Integer limit, Locale locale) {
-    List<OrderWideListDto> result = orderDao.getMyOrdersWithState(email, currencyPair, status, operationType, null, offset, limit, locale);
+    List<OrderWideListDto> result = orderDao.getMyOrdersWithState(userService.getIdByEmail(email), currencyPair, status, operationType, SCOPE, offset, limit, locale);
 
     return result;
   }
@@ -1304,9 +1328,9 @@ public class OrderServiceImpl implements OrderService {
   @Transactional(readOnly = true)
   @Override
   public List<OrderWideListDto> getMyOrdersWithState(String email, CurrencyPair currencyPair, OrderStatus status,
-                                                     OperationType operationType,
+                                                     OperationType operationType, String scope,
                                                      Integer offset, Integer limit, Locale locale) {
-    return orderDao.getMyOrdersWithState(email, currencyPair, status, operationType, null, offset, limit, locale);
+    return orderDao.getMyOrdersWithState(userService.getIdByEmail(email), currencyPair, status, operationType, scope, offset, limit, locale);
   }
 
   @Override
@@ -1314,7 +1338,7 @@ public class OrderServiceImpl implements OrderService {
   public List<OrderWideListDto> getMyOrdersWithState(String email, CurrencyPair currencyPair, List<OrderStatus> statuses,
                                                      OperationType operationType,
                                                      Integer offset, Integer limit, Locale locale) {
-    return orderDao.getMyOrdersWithState(email, currencyPair, statuses, operationType, null, offset, limit, locale);
+    return orderDao.getMyOrdersWithState(userService.getIdByEmail(email), currencyPair, statuses, operationType, null, offset, limit, locale);
   }
 
 
@@ -1570,7 +1594,6 @@ public class OrderServiceImpl implements OrderService {
       arrayList.add(candle.getBaseVolume());
       arrayListMain.add(arrayList);
     }
-
     try {
       return objectMapper.writeValueAsString(new OrdersListWrapper(arrayListMain,
               backDealInterval.getInterval(), currencyPairId));
@@ -1614,6 +1637,23 @@ public class OrderServiceImpl implements OrderService {
   public List<OrdersCommissionSummaryDto> getOrderCommissionsByPairsForPeriod(LocalDateTime startTime, LocalDateTime endTime,
                                                                               List<Integer> userRoleIdList) {
     return orderDao.getOrderCommissionsByPairsForPeriod(startTime, endTime, userRoleIdList);
+  }
+
+  //wolper 23.04.18
+  //Returns the list of the latest exchange rates for each currency to USD
+  @Override
+  @Transactional(readOnly = true)
+  public Map<Integer, RatesUSDForReportDto> getRatesToUSDForReport() {
+    return orderDao.getRatesToUSDForReport().stream().collect(Collectors.toMap(RatesUSDForReportDto::getId, Function.identity()));
+  }
+
+
+  //wolper 24.04.18
+  //Returns the list of the latest exchange rates for each currency to USD
+  @Override
+  @Transactional(readOnly = true)
+  public Map<String, RatesUSDForReportDto> getRatesToUSDForReportByCurName() {
+    return orderDao.getRatesToUSDForReport().stream().collect(Collectors.toMap(RatesUSDForReportDto::getName, Function.identity()));
   }
 }
